Support relative files for in and out
